requirements.txt
beautifulsoup4==4.10.0
celery==5.2.6
Django==3.2.6
djangorestframework==3.13.1
django-ckeditor==6.4.0
facebook-page-scraper==0.1.8
google-searching==0.8
goose3==3.1.11
matplotlib==3.5.1
newspaper3k==0.2.8
numpy==1.22.1
pandas==1.3.5
Pillow==9.0.1
pytrends==4.7.3
requests==2.27.1
yake==0.4.8
wordcloud==1.8.1